Fs_changes
I modified the sample rate of both the DAC and ADC. I also added new parameters into run_chirp.py so there is a different value for Fs_DAC and Fs_ADC as they are not the exact same value. I tested these codes and was successfully able to produce the correct frequency signals and receive them correctly in the plots.